Split iOS cloud sync runner
